inital porting

+++ MachineLearning.Benchmarks/GlobalUsings.cs
@@ -1,6 +1,6 @@
 ﻿global using Ametrin.Numerics;
 global using BenchmarkDotNet.Attributes;
 global using System.Runtime.InteropServices;
-global using SimdVector = System.Numerics.Vector<double>;
+global using SimdVector = System.Numerics.Vector<float>;
 global using SimdVectorHelper = System.Numerics.Vector;
-global using Weight = double;
+global using Weight = float;
