+++ MachineLearning.Model/GlobalUsings.cs
@@ -1,6 +1,6 @@
 ﻿global using Ametrin.Numerics;
 global using Ametrin.Optional;
 global using Ametrin.Utils;
-global using Weight = double;
-global using SimdVector = System.Numerics.Vector<double>;
+global using Weight = float;
+global using SimdVector = System.Numerics.Vector<float>;
 global using SimdVectorHelper = System.Numerics.Vector;
