+++ MachineLearning.Training/GlobalUsings.cs
@@ -3,7 +3,7 @@ global using Ametrin.Utils.Optional;
 global using Ametrin.Utils;
 global using MachineLearning.Model;
 global using System.Diagnostics;
-global using Weight = double;
+global using Weight = float;
 
 using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("MachineLearning.Benchmarks")]
